+++ net/netmon/interfaces_default_route_test.go
@@ -3,7 +3,7 @@
 
 //go:build linux || (darwin && !ts_macext)
 
-package interfaces
+package netmon
 
 import (
 	"testing"
