+++ net/netns/netns_darwin.go
@@ -18,7 +18,6 @@ import (
 	"golang.org/x/net/route"
 	"golang.org/x/sys/unix"
 	"tailscale.com/envknob"
-	"tailscale.com/net/interfaces"
 	"tailscale.com/net/netmon"
 	"tailscale.com/net/tsaddr"
 	"tailscale.com/types/logger"
@@ -62,12 +61,12 @@ func getInterfaceIndex(logf logger.Logf, netMon *netmon.Monitor, address string)
 	// Helper so we can log errors.
 	defaultIdx := func() (int, error) {
 		if netMon == nil {
-			idx, err := interfaces.DefaultRouteInterfaceIndex()
+			idx, err := netmon.DefaultRouteInterfaceIndex()
 			if err != nil {
 				// It's somewhat common for there to be no default gateway route
 				// (e.g. on a phone with no connectivity), don't log those errors
 				// since they are expected.
-				if !errors.Is(err, interfaces.ErrNoGatewayIndexFound) {
+				if !errors.Is(err, netmon.ErrNoGatewayIndexFound) {
 					logf("[unexpected] netns: DefaultRouteInterfaceIndex: %v", err)
 				}
 				return -1, err
