+++ net/routetable/routetable_bsd.go
@@ -17,7 +17,7 @@ import (
 
 	"golang.org/x/net/route"
 	"golang.org/x/sys/unix"
-	"tailscale.com/net/interfaces"
+	"tailscale.com/net/netmon"
 	"tailscale.com/types/logger"
 )
 
@@ -80,7 +80,7 @@ func (r RouteEntryBSD) Format(f fmt.State, verb rune) {
 
 // ipFromRMAddr returns a netip.Addr converted from one of the
 // route.Inet{4,6}Addr types.
-func ipFromRMAddr(ifs map[int]interfaces.Interface, addr any) netip.Addr {
+func ipFromRMAddr(ifs map[int]netmon.Interface, addr any) netip.Addr {
 	switch v := addr.(type) {
 	case *route.Inet4Addr:
 		return netip.AddrFrom4(v.IP)
@@ -102,7 +102,7 @@ func ipFromRMAddr(ifs map[int]interfaces.Interface, addr any) netip.Addr {
 }
 
 // populateGateway populates gateway fields on a RouteEntry/RouteEntryBSD.
-func populateGateway(re *RouteEntry, reSys *RouteEntryBSD, ifs map[int]interfaces.Interface, addr any) {
+func populateGateway(re *RouteEntry, reSys *RouteEntryBSD, ifs map[int]netmon.Interface, addr any) {
 	// If the address type has a valid IP, use that.
 	if ip := ipFromRMAddr(ifs, addr); ip.IsValid() {
 		re.Gateway = ip
@@ -128,7 +128,7 @@ func populateGateway(re *RouteEntry, reSys *RouteEntryBSD, ifs map[int]interface
 
 // populateDestination populates the 'Dst' field on a RouteEntry based on the
 // RouteMessage's destination and netmask fields.
-func populateDestination(re *RouteEntry, ifs map[int]interfaces.Interface, rm *route.RouteMessage) {
+func populateDestination(re *RouteEntry, ifs map[int]netmon.Interface, rm *route.RouteMessage) {
 	dst := rm.Addrs[unix.RTAX_DST]
 	if dst == nil {
 		return
@@ -194,7 +194,7 @@ func populateDestination(re *RouteEntry, ifs map[int]interfaces.Interface, rm *r
 
 // routeEntryFromMsg returns a RouteEntry from a single route.Message
 // returned by the operating system.
-func routeEntryFromMsg(ifsByIdx map[int]interfaces.Interface, msg route.Message) (RouteEntry, bool) {
+func routeEntryFromMsg(ifsByIdx map[int]netmon.Interface, msg route.Message) (RouteEntry, bool) {
 	rm, ok := msg.(*route.RouteMessage)
 	if !ok {
 		return RouteEntry{}, false
@@ -260,12 +260,12 @@ func routeEntryFromMsg(ifsByIdx map[int]interfaces.Interface, msg route.Message)
 func Get(max int) ([]RouteEntry, error) {
 	// Fetching the list of interfaces can race with fetching our route
 	// table, but we do it anyway since it's helpful for debugging.
-	ifs, err := interfaces.GetList()
+	ifs, err := netmon.GetInterfaceList()
 	if err != nil {
 		return nil, err
 	}
 
-	ifsByIdx := make(map[int]interfaces.Interface)
+	ifsByIdx := make(map[int]netmon.Interface)
 	for _, iif := range ifs {
 		ifsByIdx[iif.Index] = iif
 	}
