+++ tstest/integration/tailscaled_deps_test_darwin.go
@@ -27,7 +27,6 @@ import (
 	_ "tailscale.com/logtail"
 	_ "tailscale.com/net/dns"
 	_ "tailscale.com/net/dnsfallback"
-	_ "tailscale.com/net/interfaces"
 	_ "tailscale.com/net/netmon"
 	_ "tailscale.com/net/netns"
 	_ "tailscale.com/net/proxymux"
