+++ tstest/integration/vms/derive_bindhost_test.go
@@ -8,19 +8,19 @@ import (
 	"runtime"
 	"testing"
 
-	"tailscale.com/net/interfaces"
+	"tailscale.com/net/netmon"
 )
 
 func deriveBindhost(t *testing.T) string {
 	t.Helper()
 
-	ifName, err := interfaces.DefaultRouteInterface()
+	ifName, err := netmon.DefaultRouteInterface()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	var ret string
-	err = interfaces.ForeachInterfaceAddress(func(i interfaces.Interface, prefix netip.Prefix) {
+	err = netmon.ForeachInterfaceAddress(func(i netmon.Interface, prefix netip.Prefix) {
 		if ret != "" || i.Name != ifName {
 			return
 		}
